pyproject.toml
[project]
name = "ai-systems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen>=0.9.7",
    "beautifulsoup4==4.13.4",
    "camel==0.1.2",
    "celery==5.5.3",
    "colorama==0.4.6",
    "dashscope==1.23.8",
    "duckduckgo_search==6.4.2",
    "elasticsearch==9.0.2",
    "fastapi==0.135.1",
    "fire==0.7.0",
    "jieba==0.42.1",
    "jieba3k==0.35.1",
    "langchain==0.3.27",
    "langchain_community==0.3.27",
    "langchain_core==1.2.17",
    "langchain_experimental==0.3.4",
    "langchain_mcp_adapters==0.1.9",
    "langchain_mcp_tools==0.2.4",
    "langchain_openai==0.3.28",
    "langchain_text_splitters==0.3.9",
    "langgraph==1.0.10",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.9",
    "matplotlib==3.10.3",
    "mcp",
    "openai==1.97.1",
    "pydantic",
    "pymilvus==2.5.12",
    "python-dotenv==1.1.1",
    "python_magic==0.4.27",
    "rank_bm25",
    "Requests==2.32.5",
    "tabulate==0.9.0",
    "typing_extensions",
    "volcengine_python_sdk==3.0.15",
    "mem0ai==1.0.5",
    "posthog",
    "python-dateutil",
    "pytz",
    "qdrant-client",
    "protobuf",
    "grpcio",
    "opentelemetry-exporter-otlp",
    "opentelemetry-exporter-otlp-proto-common",
    "opentelemetry-proto",
    "googleapis-common-protos",
    "chromadb",
    "neo4j",
    "numpy",
    "pydantic_core==2.41.5",
    "typing_inspection",
    "overrides",
    "tenacity",
    "pybase64",
    "annotated_types",
    "jsonschema",
    "attrs",
    "referencing",
    "rpds-py",
    "jsonschema-specifications",
    "httpx",
    "idna",
    "orjson",
    "pyyaml",
    "opentelemetry-api",
    "importlib-metadata",
    "opentelemetry-sdk",
    "opentelemetry-semantic-conventions",
    "opentelemetry-exporter-otlp",
    "opentelemetry-exporter-otlp-proto-grpc",
    "packaging",
    "langsmith==0.7.14",
    "urllib3",
    "certifi",
    "requests-toolbelt",
    "xxhash",
    "uuid-utils",
    "jsonpatch",
    "jsonpointer",
    "sniffio",
    "anyio",
    "distro",
    "tiktoken",
    "six",
    "sqlalchemy==2.0.48",
    "starlette==0.52.1",
    "httpcore",
    "h11",
    "socksio",
    "jiter",
]
